Add harp voicing engine fixes, ABC parser, chord display, and MIDI generator
- Fix midi_to_harp_string/harp_string_to_midi for non-C keys (key-relative octave)
- Add voice_from_satb SATB voicing engine with R1-R8 scoring rules
- Add snap_to_diatonic, identify_chord dyad handling, 29-string mode check
- Add abc.rs: full ABC parser with duration, key sig, tempo, SATB/lead sheet
- Add chord_display.rs: traditional, spoken, example, interval formatters
- Add harp-midi crate: batch CLI generates MIDI for all hymns × 7 church modes
  Format 1, 4 tracks (conductor, melody, RH harp, LH harp), GM Harp instrument
- 53 tests in harp-voicing-wasm, all passing

+++ harp-voicing-wasm/src/harp.rs
@@ -1,7 +1,7 @@
 use serde::Serialize;
 
 /// Maximum diatonic span per hand (10 strings).
-pub const MAX_HAND_SPAN: u8 = 10;
+pub const MAX_HAND_SPAN: i16 = 10;
 
 /// Maximum fingers per hand.
 pub const FINGERS_PER_HAND: usize = 4;
@@ -20,7 +20,7 @@ pub enum Hand {
 #[derive(Debug, Clone, Serialize)]
 pub struct FingerNote {
     pub midi: u8,
-    pub string_num: u8,
+    pub string_num: i16,
     pub finger: u8, // 1=thumb, 2=index, 3=middle, 4=ring
     pub hand: Hand,
     pub note_name: String,
